Delay sending the fake mouse move
This is so a user can get at a control that would normally be hidden by
a mouseover behavior. This is to deal with a control that displays a
fast pop-up grid when hovering over the name, completely obscuring the
control. With this change, the user has to hover over it for one second
before the hover behavior is activated.