Better handle default values in test steps
This commit alters the handling of default values for Name, AlwaysRun, and
ExitBranchOnError. The latter two were virtual properties before, and they
were overridden in CleanupTestCase and ActualTestCase to change their
DefaultValueAttributes. After some experiments with the XmlSerializer, it
turns out that's a very bad way to do it.

It turns out the PropertyGrid has a dynamic way of learning whether the
value of the property is default or not-- ShouldSerialize methods,
described here: http://msdn.microsoft.com/en-us/library/53b8022e.aspx

With this, and by requiring the test step's default name as a parameter to
the constructor, we can finally show only those values that have changed.
The XmlSerializer looks like it respects these methods, too.