Convert save file format to XML
This commit assigns element names and namespaces to each of the test
steps and DLLs. It also adds the PreferredXmlPrefixAttribute, which lets
an assembly declare XML prefixes it would like to have. This prevents
the XmlSerializer from assigning prefixes automatically, which would
make it much harder to diff the resulting files.