Get rid of individual UI class types by implementing pattern interfaces directly
This implements UI automation patterns directly through the
GetControlInterface<T>() method. Suddenly we don't need the individual
control types, and we have a more consistent and dynamic way of talking to
a kind of control.

One thing this does eliminate is the properties that would show up in the
type descriptor; we want these to show up as attributes in our XPath
implementation, so they'll make a comeback as dynamic properties.