Add XPath support for UIA controls
This commit adds a custom XPath parser and a uia: prefix on code paths. The
framework then interprets element names as control types and uses XPath
predicates to search for the right controls, rather than the limited
indexers that we were using before. This opens up additional functions and
operators that can be used to test controls, such as starts-with() to see
if the name of a window starts with (not just equals) the right name.

There will probably need to be more adjustments to the generation of code
paths in the CodePath property. I also think there's a bug waiting to be
found in using Index instead of IDIndex for identification; someone's going
to have to detect that Index isn't initialized for controls discovered deep
in the tree.