Let's get precise: Make sure clicks happen on target point
Final clicks don't always happen at the right spot. This is because the raw
input code leaves out the MOUSEEVENTF_MOVE flag when sending clicks; though
it specifies a point, the point is being ignored from the click and depends
on the previous mouse move ending correctly. The reliability of clicks has
improved dramatically.