The QAliber CommonTestCases DLL-- in fact, all add-ins-- should only go to User Assemblies
Testing highlighted a problem which arises when two copies of the same
add-in exist, one in "User Assemblies" and the other in the program
directory-- the Clone process and the XML readers and writers get confused
as to which DLLs they're using, and so the same types end up getting
duplicated in memory and failing comparison tests with themselves and...
it's just a mess. So user assemblies *only* go to User Assemblies.