Rewrite the MultipleSelectionTypeEditor to fetch list dynamically
There are two places where the items to be displayed in a drop-down are
stored in the property itself, presumably so the UITypeEditor can find
them. This is one place-- the fix here is to use the property's name to
look up a method that can provide the list items. No more saving it to XML,
or having to worry about the list being up-to-date.