Support variable selection with type converters, not UI editors
This commit solves a few problems. The main one is that the list of
variables to display was being saved in the property itself; this is ugly,
and it means that the list is saved to XML and cloned around everywhere.
The other problem was that there was a UI editor being used for a standard
items drop-down, something a TypeConverter can do.

So this commit replaces the UI editor and the variable list properties with
type converters and simple string properties.