Replace the binary serializer Clone with a faster MemberwiseClone solution
I recently got an OutOfMemoryException working with QAliber. While trying
to profile the program, I noticed a lot of time and a lot of memory was
being spent on clone operations, which are used to support drag-and-drop,
copy and paste, and undo operations in the GUI. I'm sure tons of nasty
things were being duplicated, including probably the icons for each node,
so I decided we should have an ordinary in-place memberwise clone.

The Clone overrides are there to either (a) clone some child object that
shouldn't be shared, or (b) clear the output properties in the resulting
clone. Most cases of (b) aren't strictly necessary, as that wasn't being
done before anyways, but it does make the result a little nicer.

We didn't need to hold on to the cpuCounter in LogResources-- it was being
re-created every time anyways.