Rewrite some of the attributes on test steps
Instead of inventing a category for every type of control, we put input
properties in "Behavior" and outputs in "Results", both of which sort
before the "Test" categories made by the base class. I also remove the
numbers on DisplayNames (properties are now sorted by appearance, not
name), fix descriptions to sentence-style, and add default values where
appropriate.

There are a lot more to fix, but this is a start.