Don't copy add-in's dependencies to User Assemblies
This is not a complete fix, just a stop-gap measure. Here's what it is.

Originally, we were copying all dependencies to User Assemblies. This again
meant that we had duplicate assemblies, and method lookups were failing as
a result. Not copying all the assemblies to User Assemblies meants that
only the add-in itself is there, and all of its dependencies can be loaded
from Binaries like normal.

The problem that's left is that some of CommonTestCases' dependencies are
not dependencies of the main app, such as QAliber.ImageHandling. That is
built to Binaries, but it should be in User Assemblies. What we need is to
move the project's build areas around so that only the main app's
dependencies end up next to it, and the User Assemblies can be free to keep
their own.

Another possible solution for this might be to somehow use PrivateBinPath
and Assembly.Load instead of Assembly.LoadFrom so that all assemblies in
the app can use the same context and assembly-finding mechanism.