Add value and grid patterns to UIAControl
...and actually the grid implementation is incomplete, because this commit
focuses on an implementation that handles DataGridView controls.
DataGridViews are particularly nasty because they don't implement UIA at
all, so the implementation here tries to mask that by doing string tests on
the cell names that come back and arranging those into a grid. This also
tries to compensate for fast-changing grids by going back and recapturing
if it thinks it got an incomplete grid.