Show nifty currently-playing icons when selecting "Play current test"
For the longest time, I couldn't figure out why this didn't work. Turns out
there's a field in the ExecutionContainer which needs to be set to the
currently playing scenario. It's set when the master play button is hit,
but not for the context menu play item.

I couldn't figure out a decent way to get that info back to the
ExecutionContainer. They already had a static event on the ScenarioControl,
so I figured a second couldn't hurt.