Fix Ctrl+C for children of the LogViewerControl
LogViewerControl was stealing Ctrl+C to try to do a copy operation on the
tree view, which didn't work anyways as it expected the copy command to be
invoked from the tree's context menu. Now we only grab Ctrl+C if the tree
view has focus, letting other controls have working copy operations. If the
tree view does have focus, we copy the currently selected node.