Revamp the Logging class design
In advance of doing truly automated testing, I figured the logging system
could use some improvements. Here, I'm removing the singleton logger
pattern (I hate singletons!) in favor of static methods that invoke methods
on a Log instance that can be swapped out. (Alternatively, I could have
made the whole class static.)

This also stops accepting style attributes for the log entries, which
nobody was using, and which I'm not sure were a good idea in the first
place.

Next, I will improve the design of the XML in the log files, so that
someone could have the hope to write XSL for it.