Start the variable rewrite
The idea here was to remove some of the global state-- most especially the
current log object, the run state (currently in TestCase), and the
scenario variables so we can have more control over them. I tried to move
the log, but that's a bear-- the UIA engine, among other things, relies on
there being a global current log.

Also, the remoting stuff is gone now, but that's fine, we can stand to
redesign it if we ever need it again.